+++ ring-buffer/inc/ring-buffer.h
@@ -81,6 +81,16 @@ typedef struct {
     void * data;
 } RingBufferInterface;
 
+/**
+ * @brief Enum with all the possible return codes for the ring buffer functions
+ */
+typedef enum {
+    RING_BUFFER_OK,
+    RING_BUFFER_NULL_POINTER,
+    RING_BUFFER_EMPTY,
+    RING_BUFFER_FULL
+} RingBufferReturnCode;
+
 
 /**
  * @brief Initialize the buffer
@@ -92,6 +102,9 @@ typedef struct {
  * @param capacity The maximum number of elements of the buffer
  * @param cs_enter A pointer to a function that should manage a critical section (can be NULL)
  * @param cs_exit A pointer to a function that should exit a critical section (can be NULL)
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler is NULL
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_init(buffer, type, capacity, cs_enter, cs_exit) _ring_buffer_init((RingBufferInterface *)(buffer), sizeof(type), capacity, cs_enter, cs_exit)
 
@@ -124,7 +137,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param item A pointer to the item to insert
- * @return bool True if the item is inserted correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler or the item are NULL
+ *     - RING_BUFFER_FULL if the buffer is full
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_push_front(buffer, item) _ring_buffer_push_front((RingBufferInterface *)(buffer), (void *)(item))
 
@@ -133,7 +149,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param item A pointer to the item to insert
- * @return bool True if the item is inserted correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler or the item are NULL
+ *     - RING_BUFFER_FULL if the buffer is full
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_push_back(buffer, item) _ring_buffer_push_back((RingBufferInterface *)(buffer), (void *)(item))
 
@@ -143,7 +162,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param out A pointer to a variable where the removed item is copied into
- * @return bool True if the item is removed correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler is NULL
+ *     - RING_BUFFER_EMPTY if the buffer is empty
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_pop_front(buffer, out) _ring_buffer_pop_front((RingBufferInterface *)(buffer), (void *)(out))
 
@@ -153,7 +175,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param out A pointer to a variable where the removed item is copied into
- * @return bool True if the item is removed correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler is NULL
+ *     - RING_BUFFER_EMPTY if the buffer is empty
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_pop_back(buffer, out) _ring_buffer_pop_back((RingBufferInterface *)(buffer), (void *)(out))
 
@@ -162,7 +187,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param out A pointer to a variable where the item is copied into
- * @return bool True if the item is copied correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler or out are NULL
+ *     - RING_BUFFER_EMPTY if the buffer is empty
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_front(buffer, out) _ring_buffer_front((RingBufferInterface *)(buffer), (void *)(out))
 
@@ -171,7 +199,10 @@ typedef struct {
  *
  * @param buffer The buffer handler structure
  * @param out A pointer to a variable where the item is copied into
- * @return bool True if the item is copied correctly, false otherwise
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler or out are NULL
+ *     - RING_BUFFER_EMPTY if the buffer is empty
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_back(buffer, out) _ring_buffer_back((RingBufferInterface *)(buffer), (void *)(out))
 
@@ -200,6 +231,9 @@ typedef struct {
  * @details The actual data is not erased, only the size is modified
  *
  * @param buffer The buffer handler structure
+ * @return RingBufferReturnCode
+ *     - RING_BUFFER_NULL if the buffer handler is NULL
+ *     - RING_BUFFER_OK otherwise
  */
 #define ring_buffer_clear(buffer) _ring_buffer_clear((RingBufferInterface *)(buffer))
 
@@ -208,7 +242,7 @@ typedef struct {
 /*         USE THE MACRO INSTEAD          */
 /******************************************/
 
-void _ring_buffer_init(
+RingBufferReturnCode _ring_buffer_init(
     RingBufferInterface * buffer,
     size_t data_size,
     size_t capacity,
@@ -218,15 +252,15 @@ void _ring_buffer_init(
 bool _ring_buffer_is_empty(RingBufferInterface * buffer);
 bool _ring_buffer_is_full(RingBufferInterface * buffer);
 size_t _ring_buffer_size(RingBufferInterface * buffer);
-bool _ring_buffer_push_front(RingBufferInterface * buffer, void * item);
-bool _ring_buffer_push_back(RingBufferInterface * buffer, void * item);
-bool _ring_buffer_pop_front(RingBufferInterface * buffer, void * out);
-bool _ring_buffer_pop_back(RingBufferInterface * buffer, void * out);
-bool _ring_buffer_front(RingBufferInterface * buffer, void * out);
-bool _ring_buffer_back(RingBufferInterface * buffer, void * out);
+RingBufferReturnCode _ring_buffer_push_front(RingBufferInterface * buffer, void * item);
+RingBufferReturnCode _ring_buffer_push_back(RingBufferInterface * buffer, void * item);
+RingBufferReturnCode _ring_buffer_pop_front(RingBufferInterface * buffer, void * out);
+RingBufferReturnCode _ring_buffer_pop_back(RingBufferInterface * buffer, void * out);
+RingBufferReturnCode _ring_buffer_front(RingBufferInterface * buffer, void * out);
+RingBufferReturnCode _ring_buffer_back(RingBufferInterface * buffer, void * out);
 void * _ring_buffer_peek_front(RingBufferInterface * buffer);
 void * _ring_buffer_peek_back(RingBufferInterface * buffer);
-void _ring_buffer_clear(RingBufferInterface * buffer);
+RingBufferReturnCode _ring_buffer_clear(RingBufferInterface * buffer);
 
 // Function that substitute cs_enter and cs_exit if they are NULL
 void _ring_buffer_cs_dummy(void);